Se crea api para login, se modifica conexion para uso global
